% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.prob.pnb.R
\name{drop.prob.pnb}
\alias{drop.prob.pnb}
\title{Calculate dropout probability for PNB model}
\usage{
drop.prob.pnb(pars, s, y, offset, max.val = 1e+100, tol = 1e-10)
}
\arguments{
\item{pars}{vector containing parameters, gamma, alpha, and delta}

\item{s}{PNB mixture parameter}

\item{y}{vector containing observed expression values}

\item{offset}{vector containing log effective library size}

\item{max.val}{maximum value for computation}

\item{tol}{tolerance for computation}
}
\value{
a vector containing dropout probability
}
\description{
Function to calculate dropout probability
}
