% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung}
\alias{lung}
\title{Lung scRNA-seq data}
\format{a data frame with 60675 genes (rows) and 145 samples (cols)}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE96106}
}
\usage{
lung
}
\description{
A dataset containing the raw count expression of lung epithelium scRNA-seq
}
\keyword{datasets}
