\name{scdoc.md2html}
\alias{scdoc.md2html}
\title{Convert an md file to HTML/PDF}
\usage{
  scdoc.md2html(inputFile, outputFile)
}
\arguments{
  \item{inputFile}{input file name (md file)}

  \item{outputFile}{output file name (html file)}
}
\description{
  Convert an md file to HTML/PDF
}
\details{
  the output pdf file name will be derivated from
  \code{outputFile}.
}

