\name{scdoc}
\alias{scdoc}
\title{Initialize scdoc or knit and convert a Rmd file using scdoc}
\usage{
  scdoc(file = NULL)
}
\arguments{
  \item{file}{file to knit and convert using scdoc, if NULL
  initialize scdoc in R Studio}
}
\description{
  Initialize scdoc or knit and convert a Rmd file using
  scdoc
}
\details{
  Use the command \code{scdoc()} to define scdoc() as the
  command to apply when clicking on \strong{Knit HTML}
  button in R Studio. Alternativaly, you can also specify a
  file to knit and convert into HTML/PDF.
}

