% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish}
\alias{zebrafish}
\title{Zebrafish scRNA-seq data}
\format{a data frame with 21876 genes (rows) and 246 samples (cols)}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE100911}
}
\usage{
zebrafish
}
\description{
A dataset containing the raw count expression of zebrafish scRNA-seq
}
\keyword{datasets}
