\name{zotxt2json}
\alias{zotxt2json}
\title{Extract references from a Rmd/md file with Zotxt and create a JSON file}
\usage{
  zotxt2json(file, json = NULL, return.file.name = FALSE)
}
\arguments{
  \item{file}{file to parse}

  \item{json}{name of the JSON file to produce (if NULL
  automatically determined)}

  \item{return.file.name}{if TRUE, return the name of the
  JSON file}
}
\description{
  Extract references from a Rmd/md file with Zotxt and
  create a JSON file
}
\note{
  Zotero should be opened for this function to work.
}

