\name{zotxt2yaml}
\alias{zotxt2yaml}
\title{Extract references from a Rmd/md file with Zotxt and convert it to YAML}
\usage{
  zotxt2yaml(file, keep = FALSE, display = TRUE)
}
\arguments{
  \item{file}{file to parse}

  \item{keep}{if TRUE, keep the json and YAML files
  produced}

  \item{display}{if TRUE, display the result in the
  console}
}
\description{
  Extract references from a Rmd/md file with Zotxt and
  convert it to YAML
}
\note{
  Zotero should be opened for this function to work.
}

